// srm_ui/src/app/main/Models/etendering/ViewModels/rfq-supplier-partline-attachment-view-model.ts
export class RFQSupplierPartLineAttachmentViewModel {
    rFQPartLineAttachmentId : string;
    rFQSupplierPartLineId: string;
    rFQId: string;
    supplierId: string;
    attachmentId: string;
    comment: string;
    surrogateById: string;
    surrogationDate: string;
    isDeleteandDownloadEnabledVisible:boolean;
    srNo:string;
    title : string;
    fileName : string;
    type : string;
    buyFileName:string;
    buyETMediaId : any;
    buyFileExtension:string;
    rfqSupplierPartLineId: string;
    
}